ai-service/requirements.txt
flask==3.0.3
flask-limiter==3.7.0
flask-cors==4.0.1
groq==0.9.0
python-dotenv==1.0.1
redis==5.0.8
pytest==8.2.2
requests==2.31.0
sentence-transformers==2.7.0
httpx==0.27.0